Ensure the language code from a speech sequence is normalized before checking if the synth supports it. (#18541)
Summary of the issue:

Recently, NVDA gained the ability to report the name of a language, when speaking text in that language and the synthesizer did not support it. However, in particular situations such as reading mathML, NvDA would report a language was not supported when it was. E.g. reading MathML in English with eSpeak, NvDA would report "English (not supported)".
The reason for this seems to be that a non-normalized language code E.g. en-gb was being compared against normalized language codes, such as en_GB, and never matching.
Description of user facing changes:

NVDA will no longer incorrectly report a particular language is not supported.
Description of developer facing changes:
Description of development approach:

In speech.languageHandling.getLangToReport: normalize the language.