Add-on store: Fix column sorting combo-box when clicking add-on list column header (#18547)
Closes #18540
Summary of the issue:

When clicking an add-on's list header in the add-on store, the list is sorted, but the combo-box sorting indication is not updated accordingly.
Description of user facing changes:

Sorting order combo-box will be sorted accordingly.
Description of developer facing changes:

None
Description of development approach:

Set the value of the combo-box when a header is clicked.