QuickstartLinks
Added links (like a mini table of contents) to the Quick start guide.  I often refer people to this section when just starting out with NVDA, and while in some cases they may be at a particular point (Having already installed NVDA, or needing to acquire it), in others, it is unclear, so I wanted to make it easy to point them to the start of the Quick Start Guide but for them to readily be able to jump to the part they need.