Disable auto merge for snapshot deployments (#18553)
Fixup of #18244
Summary of the issue:

Beta/rc snapshots will fail to deploy if there are merge conflicts between master and the snapshot commit.
This will cause releases to be blocked, as we expect a passing beta/rc snapshot to perform the tagged release.
Beta/rc don't need to have merge conflicts resolved to be a valid build, though regular merging to minimize these is ideal.