Add links to the start of the Quick Start Guide (#18550)
Summary of the issue:
I often refer people to this section when just starting out with NVDA,
and while in some cases they may be at a particular point (Having
already installed NVDA, or needing to acquire it), in others, it is
unclear, so I wanted to make it easy to point them to the start of the
Quick Start Guide but for them to readily be able to jump to the part
they need.

Description of user facing changes:
I added links (like a mini table of contents) to the start of the Quick
start guide, to point to the anchors which already existed in the
various headings in that section.