Check if all tests pass as a single job (#18554)
Part of #17878 

Summary of the issue:
GitHub Rulesets require specifying every job you wish to block a
PR/push.
With many rulesets, adding new blocking jobs is fragile.
For example, if a new system test suite is added, it will need to be
added to every ruleset.

Description of user facing changes:
none

Description of developer facing changes:
Blocking tests are added to the job `allTestsPass`.

Description of development approach:
We can use that job to confirm if a build can be merged or not via
rulesets.

Testing strategy:
actions 

Known issues with pull request:
none