Add an "Install date" column in the Add-on Store (#18560)
In the Add-on Store, the install date of an add-on is available in the Details pane, and only for add-ons installed from the Store.
Having the install date displayed in a column of the list would allow to view them all at once and to sort the add-on by install date. This can be very useful to track a bug recently appeared.