Move builds to windows-2025 (#18592)
Fixes #15104
Fixes #14372
Fixes #14353
Summary of the issue:

    Most NVDA developers are on Windows 11, including all of NV Access, but NVDA is still being built on Windows Server 2022 (Windows 10 based).
    Windows 11 is the expected environment for system testing and builds.
    GitHub actions will move windows-latest to windows-2025 this year.
    A bug exists with windows 10 notepad that prevents NVDA from focusing it. This causes our symbol pronunciation tests to be disabled.

Description of user facing changes:

None
Description of developer facing changes:

symbols tests are re-enabled
Description of development approach:

    Set windows-2025 rather than windows-latest for build steps
    Explicitly call setup-python more frequently as our required version of python doesn't come pre-installed in windows-2025. This can be fixed in 

Update to python 3.13 #18591
explicitly install the windows-sdk when creating symbols, as it appears that symstore.exe is not installed at all or not installed in the expected place.
Enable symbol pronunciation tests again.