Parallelize GitHub actions builds for 32/64 bit (#18596)
Part of #18207
Summary of the issue:

While developing 64bit, we want to be able to build 32 and 64 bit versions of NVDA simultaneously, until a stable version is ready.
Similarly, it would be helpful to do the same for python upgrades.
Description of user facing changes:

None
Description of developer facing changes:

None
Description of development approach:

Create dual builds for the following jobs:

    Build NVDA
    Create launcher
    Unit tests
    System tests
    Creating symbols