Encapsulate registry keys more consistently (#18608)
Prep work for #16304

Summary of the issue:
Registry keys are not consistently encapsulated in the NVDA code base.
When migrating to 64bit, we need to be able to fetch legacy 32bit
versions of registry keys from the 64bit version.
Having more clear encapsulation of registry keys will make this easier.

Description of user facing changes:
None

Description of developer facing changes:
Deprecates `easeOfAccess.RegistryKey` for a more unified approach

Description of development approach:
- Moved all major NVDA registry keys into `config.RegistryKey`. Create
an `x86` version of this for use in the 64bit migration.
- Removed usages of the [reserved 0
parameter](https://docs.python.org/3/library/winreg.html#winreg.OpenKey).
The magic number should either be documented as a constant or not used.
it's available as `winreg.REG_OPTION_RESERVED` but this is not
documented or part of the official python API. Rather than making a
named constant, the value can only be the default value so it seems
cleaner to drop it.

Testing strategy:
Ensure nothing breaks in smoke testing

Known issues with pull request:
None