Update espeak-ng to resolve espeak-ng/espeak-ng#1031 (#18633)
espeak-ng/espeak-ng#1031.
Summary of the issue:

Espeak puts adjectival stress on "has been" in English in all cases, even when it's used as a verb. This was resolved upstream, and NVDA should take in this fix.
Description of how this pull request fixes the issue:

Updates espeak-ng to commit 3b8ef3d.