Add Uninstallation section to userGuide.md (#18643)
Summary of the issue:
Makes uninstall information easier to find and link to by providing a
heading in the user guide.

Description of user facing changes:
Information on uninstalling NVDA is tucked away at the bottom of the
command line section. I have given this its own heading, re-ordered to
describe uninstalling before the command line argument for it, and added
a note that NVDA's data in %appdata% is not automatically removed on
uninstalling NVDA.