A portable copy launched immediately after creation now correctly uses its own configuration (#18443)
### Link to issue number:
Closes #18442
Fix-up of #16937

### Summary of the issue:
When creating a portable NVDA and asking to launch it immediately after
creation, the portable copy does not use its own config, but the one of
the NVDA it had been created from.

In #16937, we propagate `--disable-addons` as well as `-c` /
`--config-path` for security reasons so that an administrator can update
from a user account using their own config folder, without risking to
use malicious add-ons in the user's configuration (see
[GHSA-727q-h8j2-6p45](https://github.com/nvaccess/nvda/security/advisories/GHSA-727q-h8j2-6p45)).

### Description of developer facing changes:
N/A
### Description of development approach:
* When creating a portable copy, do not pass parameters to the new
instance. In any case, there is no UAC, so no risk of privilege
escalation.
* Parameters are still being passed to the new instance during an
update.
* When installing a new copy, I have kept parameters of original NVDA.

### Testing strategy:
* portable creation: uses its own configuration
* portable update: uses the configuration of the previous portable,
even, for example when using `-c` flag