Improve checkPo step (#18646)
Summary of the issue:

As noted in #18645, if the uv-lock pre-commit step fails, running checkPo will also fail. When checkPo fails, it's unclear why, as the warning logs are not related.
Description of user facing changes:

None
Description of developer facing changes:

PR failures are slightly more obvious when changes to uv.lock have not been committed correctly