Add script to send ctrl+alt+del (#18629)
Closes #18105

Summary of the issue:
There's currently no script to send `control+alt+delete` to the
controlled end of a Remote Access connection, which means the only way
to do so if from the Remote Access menu..

Description of user facing changes:
Added an unassigned command to send `control+alt+delete`.

Description of developer facing changes:
None

Description of development approach:
Added a new `script_sendSAS` method to `globalCommands.GlobalCommands`.
Registered this script as a local script on the `_remoteClient`
singleton so that it can be executed when sending keys.

Testing strategy:
Ran from source. Bound the script to a gesture (`NVDA+alt+delete`).
Connected as leader and connected another computer as follower. Used the
gesture when controlling the local computer and the remote computer.

Known issues with pull request:
None