Change reference from "Comments" to "Notes" in the Elements list for Excel in the userGuide.md (#18715)
Update "Comments" to "Notes" in the elements list mention in the Excel section of the user guide
Link to issue number:
Summary of the issue:

The user guide covers accessing notes in Excel with the elements list, however calls them "comments" which is the old name (there is a different feature now called comments in Excel. What you can access in the elements list are notes, and this is reflected in the radio button for this section).
Description of user facing changes:

Change description of the elements list in Excel to talk about "notes" instead of "comments".