Create discussion and upload exe to beta/RCs #18732
Part of #18556
Summary of the issue:

Beta/RC tags do not upload the exe to the release assets, unlike stable releases.

Additionally, to make it easier to discuss issues with betas, creating a beta discussion is useful.
It also allows us to easily share each new release in slack and other places via Discussions RSS feed. The releases RSS feed includes tags, which duplicates posts to slack.

If a build fails due to system tests, you cannot rebuild it as the cache gets deleted.
Description of user facing changes:

None
Description of developer facing changes:

GitHub discussions are created for betas and RCs.
All releases are auto-posted to slack.
The installer artifact is uploaded to the release assets.