Add-on manifest: add 'changelog' key to document changes for the latest add-on release (#18672)
Partially closes #14041 

Summary of the issue:
NV Access add-on store does not show changelog for latest add-on
releases.

Description of user facing changes:
Lays the foundation to display add-on changelog if defined.

Description of developer facing changes:
Adds 'changelog' to add-on manifest key definitions and makes the vlaue
translatable.

Description of development approach:
Adds a new add-on manifest key named 'changelog' to document changes
between previous and latest add-on releases. This key is translatable
and is documented to instruct add-on authos on its usage and necessity.

Testing strategy:
Manual: requires building add-ons with "changelog" manifest key/value
defined.

Known issues with pull request:
None