Fix SAPI5 (#18675)
Fixes #18674. Fixes #18676. Fixes #18760.
Summary of the issue:

NVDA becomes unresponsive for a small period of time when speaking some text in SAPI5 voices.

#18658 leaves more audio data for the EndStream handler to be played, which causes NVDA to be frozen when the EndStream handler is feeding the remaining audio chunks.
Description of user facing changes:

The above issue should be fixed.
Description of developer facing changes:

None
Description of development approach:

Moves the feeding logic in EndStream to the dedicated speak thread, avoiding blocking the SAPI5 thread that EndStream runs on.

The main thread no longer waits for cancellation to complete in order to prevent freezing.

The dedicated speak thread for WASAPI will be stopped and restarted when WASAPI is turned on or off, in order to prevent errors accessing the player when WASAPI is turned off and the player no longer exists.