Detect and report the status bar in Visual Studio Code #18724
Closes #11064.
Summary of the issue:

When invoking the standard "report status bar" gesture in Visual Studio Code, NVDA says "no status line found".
Description of how this pull request fixes the issue:

Adds logic to detect (by role or IAccessible2 ID) the status bar, caching it when found. Also added a custom text override to move the line and column to the start for speech-friendliness.
Testing strategy:

Tested in various document types (plain text, Python) in Visual Studio Code.
Known issues with pull request:

The status bar is not cached on app launch since the tree interceptor is disabled. If this override is removed, status bar accesses are always fast. With the override, the first access of the status bar takes about a second on my machine.