Add GitHub action to search for duplicates (#18763)
Summary of the issue:
It would be handy to automate searching for duplicates as part of
triage.

Description of user facing changes:
When opening an issue, a user may be greeted with a list of potential
duplicates.


Description of developer facing changes:
None

Description of development approach:

- [use Github models in GitHub
actions](https://github.blog/ai-and-ml/generative-ai/automate-your-project-with-github-models-in-actions/)
- [use ai-inference](https://github.com/actions/ai-inference)

Testing strategy:
Test on production, consider swapping out AI models as necessary

A POC that GitHub agentic workflows connected to our MCP server support
this: https://github.com/nvaccess/nvda/pull/18764

Known issues with pull request:
None