Fix touch handler (#18814)
Follow-up to #18207

Summary of the issue:
Alphas are broken due to a problem with touch handler.

Description of user facing changes:
Alphas work again.

Description of developer facing changes:
None.

Description of development approach:
Cast `TouchHandler._wca` to `LPCWSTR` when passing to `CreateWindowExW`.

This is needed because the `lpClassName` parameter of `CreateWindowExW`
can accept a class atom, but ctypes doesn't know this.
This wasn't previously a problem because ctypes didn't have type
information for `CreateWindowExW`, so happily sent the arguments.

Testing strategy:
Built and self-signed NVDA. Installed and observed that NVDA didn't
crash and my touchscreen worked.

Known issues with pull request:
None.