document deprecation utils (#18809)
Follow-up to #18571
Summary of the issue:

As part of #18571, I wrote code that made handling a large number of module-level deprecations easier. Since this code made handling module-level deprecations more systematic and auditable, the code was split out into its own module.
As we wanted to get #18571 merged to unblock other work, documentation for that utility was not added in that PR.
Description of user facing changes:

None
Description of developer facing changes:

Adds documentation on the novel strategy for handling module-level deprecations.
Description of development approach:

Wrote documentation.