Fix off by one text length in EditTextInfo (#18767)
Link to issue number:
Related to #18744 
Follow up of #18348 

Summary of the issue:
When working on #18744, I again stumbled upon the fact that the story
length reported by EditTextInfo is too high. This was discovered after
system test failures caused by #18348, i.e. reporting of an extraneous
blank at the end of the document when using NVDA review cursor. Note,
this is not a bug in #18348 and was accounted for in #18744 after the
test failure.

To reproduce
1. Open notepad classic
2. type `a`
3. Position the caret at the end of the line
4. Press nvda+delete or nvda+numpad delete

Observe that NVDA reports 50%, not 100%
Observe that `nav.makeTextInfo("caret")._getStoryLength()` reports 2,
not 1

Description of user facing changes:
None

Description of developer facing changes:
None

Description of development approach:
Fixed the off by one errors by removing the `+ 1` cases.

Testing strategy:
Tested str from this pr.

Known issues with pull request:
None