Linearize release builds (#18880)
### Link to issue number:
Fixes #18867

### Summary of the issue:
Documentation is getting built after the signing request, causing it to
not be included in the launcher.
This caused the zh_TW (last docs file) to not be included in some beta
builds

### Description of user facing changes:
None 

### Description of developer facing changes:

- Release builds may now take longer
- All builds including PR builds, snapshots, etc should safely contain
all docs files

### Description of development approach:

* Refactored SCons parallelism: Added a new `sconsCores` variable in
`setSconsArgs.ps1` to control the number of build cores, defaulting to
`--all-cores` except for debug or release builds, which use `-j1` for
easier log parsing and debugging.
* Updated workflow steps in `.github/workflows/testAndPublish.yml` to
use the new `sconsCores` variable instead of inline logic for core
selection, simplifying the build commands.
[[1]](diffhunk://#diff-0228131759fc72c25ae13bc75628d794ff96d8aadef8aa21252bd315e3279b51L105-R105)
[[2]](diffhunk://#diff-0228131759fc72c25ae13bc75628d794ff96d8aadef8aa21252bd315e3279b51L349-R351)
* Added export and logging of `sconsCores` to the environment and output
in `setSconsArgs.ps1`, improving transparency of build configuration in
CI logs.
* Removed `launcher` from the default `sconsOutTargets` in
`setSconsArgs.ps1` and added a separate workflow step to build the
launcher after other targets, preventing unsafe parallel builds of the
launcher.
[[1]](diffhunk://#diff-e627d750c249f53caef3ae05f40cd965a7e53632247af4328c3e9683884cf906L2-R13)
[[2]](diffhunk://#diff-0228131759fc72c25ae13bc75628d794ff96d8aadef8aa21252bd315e3279b51L349-R351)
* appx snippet removed as per #18799