Limit SAPI4 Vol tag value to 0~65535 (#18887)
Link to issue number:

Fixes #18866.
Summary of the issue:

IBM ViaVoice SAPI4 version is unable to process volume levels greater than 65535 in Vol tags. As a result, when speaking capital letters with prosody changes, the voice may fail to follow the current volume setting and output at its default (maximum) volume.
Description of user facing changes:

The above issue should be fixed.
Description of developer facing changes:

None.
Description of development approach:

Instead of specifying two words for volume, where the low word is for the left channel and the high word is for the right channel, specify only 0~65535 in the low word. According to the SAPI4 docs, specifying 65535 should be "the maximum monaural volume"